refactor: 스크린 이름 변경
MoreScreen -> InnerMoreScreen
AppScreen => MoreScreen
